Backend/login-service/auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
python-dotenv
psycopg2-binary
